requirements.txt
# data processing
pandas==1.5.2
numpy==1.23.5
opencv-python==4.6.0.66
tqdm==4.64.1
prefect>=2.7.0
psycopg2-binary
tabulate==0.9.0
minio==7.1.12

# plotting
matplotlib==3.6.2
seaborn==0.12.1

# other helper packages
torchsummary
PyYAML==6.0